refactor(flow): move stack to ./ui
